+++ frontend/common/base/src/commonMain/kotlin/net/matsudamper/money/frontend/common/base/notification/NotificationUsageAccessGateway.kt
@@ -1,6 +1,7 @@
 package net.matsudamper.money.frontend.common.base.notification
 
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOf
 
 public interface NotificationUsageAccessGateway {
     public fun accessStateFlow(): Flow<NotificationAccessState>
@@ -11,5 +12,14 @@ public interface NotificationUsageAccessGateway {
         public data object Granted : NotificationAccessState
 
         public data object NotGranted : NotificationAccessState
+    }
+}
+
+public object EmptyNotificationUsageAccessGateway : NotificationUsageAccessGateway {
+    override fun accessStateFlow(): Flow<NotificationUsageAccessGateway.NotificationAccessState> {
+        return flowOf(NotificationUsageAccessGateway.NotificationAccessState.NotGranted)
+    }
+
+    override fun openAccessSettings() {
     }
 }
