+++ frontend/common/base/src/commonMain/kotlin/net/matsudamper/money/frontend/common/base/notification/NotificationUsageContracts.kt
@@ -1,7 +1,5 @@
 package net.matsudamper.money.frontend.common.base.notification
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.datetime.LocalDateTime
 import net.matsudamper.money.element.MoneyUsageId
 import net.matsudamper.money.element.MoneyUsageSubCategoryId
@@ -48,44 +46,3 @@ public data class NotificationUsageDetail(
     val record: NotificationUsageRecord,
     val matched: NotificationUsageMatchedRecord?,
 )
-
-public interface NotificationUsageRepository {
-    public fun notificationsFlow(): Flow<List<NotificationUsageRecord>>
-
-    public fun unaddedMatchedNotificationsFlow(): Flow<List<NotificationUsageMatchedRecord>>
-
-    public fun notAddedNotificationsFlow(): Flow<List<NotificationUsageRecord>>
-
-    public fun addedNotificationsFlow(): Flow<List<NotificationUsageRecord>>
-
-    public fun notificationDetailFlow(notificationKey: String): Flow<NotificationUsageDetail?>
-
-    public suspend fun upsertNotification(record: NotificationUsageRecordInput): String
-
-    public suspend fun markNotificationAsAdded(notificationKey: String, moneyUsageId: MoneyUsageId?)
-}
-
-public interface NotificationUsageParser {
-    public val filterDefinition: NotificationUsageFilterDefinition
-
-    public fun parse(record: NotificationUsageRecord): NotificationUsageDraft?
-}
-
-public object EmptyNotificationUsageRepository : NotificationUsageRepository {
-    override fun notificationsFlow(): Flow<List<NotificationUsageRecord>> = flowOf(emptyList())
-
-    override fun unaddedMatchedNotificationsFlow(): Flow<List<NotificationUsageMatchedRecord>> = flowOf(emptyList())
-
-    override fun notAddedNotificationsFlow(): Flow<List<NotificationUsageRecord>> = flowOf(emptyList())
-
-    override fun addedNotificationsFlow(): Flow<List<NotificationUsageRecord>> = flowOf(emptyList())
-
-    override fun notificationDetailFlow(notificationKey: String): Flow<NotificationUsageDetail?> = flowOf(null)
-
-    override suspend fun upsertNotification(record: NotificationUsageRecordInput): String {
-        return record.notificationKey
-    }
-
-    override suspend fun markNotificationAsAdded(notificationKey: String, moneyUsageId: MoneyUsageId?) {
-    }
-}
