+++ frontend/common/base/src/commonMain/kotlin/net/matsudamper/money/frontend/common/base/notification/NotificationUsageParser.kt
@@ -1,0 +1,7 @@
+package net.matsudamper.money.frontend.common.base.notification
+
+public interface NotificationUsageParser {
+    public val filterDefinition: NotificationUsageFilterDefinition
+
+    public fun parse(record: NotificationUsageRecord): NotificationUsageDraft?
+}
