+++ frontend/common/root/src/commonMain/kotlin/net/matsudamper/money/ui/root/AddMoneyUsageScreenContainer.kt
@@ -9,7 +9,6 @@ import androidx.compose.ui.Modifier
 import kotlinx.coroutines.CoroutineScope
 import net.matsudamper.money.frontend.common.base.lifecycle.LocalScopedObjectStore
 import net.matsudamper.money.frontend.common.base.nav.user.ScreenStructure
-import net.matsudamper.money.frontend.common.base.notification.EmptyNotificationUsageRepository
 import net.matsudamper.money.frontend.common.base.notification.NotificationUsageRepository
 import net.matsudamper.money.frontend.common.ui.screen.addmoneyusage.AddMoneyUsageScreen
 import net.matsudamper.money.frontend.common.viewmodel.addmoneyusage.AddMoneyUsageScreenApi
@@ -23,8 +22,7 @@ internal fun AddMoneyUsageScreenContainer(
     windowInsets: PaddingValues,
 ) {
     val koin = LocalKoin.current
-    val notificationUsageRepository = runCatching { koin.get<NotificationUsageRepository>() }
-        .getOrElse { EmptyNotificationUsageRepository }
+    val notificationUsageRepository = koin.get<NotificationUsageRepository>()
     val viewModel = LocalScopedObjectStore.current.putOrGet(Unit) { viewModelFeature ->
         AddMoneyUsageViewModel(
             scopedObjectFeature = viewModelFeature,
