+++ frontend/common/root/src/commonMain/kotlin/net/matsudamper/money/ui/root/NotificationUsageDetailScreenContainer.kt
@@ -8,7 +8,6 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import net.matsudamper.money.frontend.common.base.lifecycle.LocalScopedObjectStore
 import net.matsudamper.money.frontend.common.base.nav.user.ScreenStructure
-import net.matsudamper.money.frontend.common.base.notification.EmptyNotificationUsageRepository
 import net.matsudamper.money.frontend.common.base.notification.NotificationUsageRepository
 import net.matsudamper.money.frontend.common.ui.screen.root.add.NotificationUsageDetailScreen
 import net.matsudamper.money.frontend.common.viewmodel.root.add.NotificationUsageDetailViewModel
@@ -21,8 +20,7 @@ internal fun NotificationUsageDetailScreenContainer(
     windowInsets: PaddingValues,
 ) {
     val koin = LocalKoin.current
-    val repository = runCatching { koin.get<NotificationUsageRepository>() }
-        .getOrElse { EmptyNotificationUsageRepository }
+    val repository = koin.get<NotificationUsageRepository>()
     val viewModel = LocalScopedObjectStore.current.putOrGet(current.notificationUsageKey) {
         NotificationUsageDetailViewModel(
             scopedObjectFeature = it,
